requirements (2).txt
gradio==3.50.2
openai==0.28
pytesseract
pillow

